fix(input): correctly parse signed PS/2 mouse deltas
Casting the 8-bit dx/dy data bytes directly to i8 works perfectly for movements under 127 pixels, because the two's complement representation is identical. However, for fast movements >127 pixels, the unsigned value wraps to a negative value when cast to i8 (e.g. 150 becomes -106), causing the cursor to jump wildly in the opposite direction.

We now properly sign-extend the 8-bit value using the dedicated X and Y sign bits (bits 4 and 5) from the flags register to form a 16-bit signed integer before passing it to the input event queue.

+++ src/input/mod.rs
@@ -170,14 +170,14 @@ pub fn inject_key_event(ascii: u8, pressed: bool) {
 /// Called by the USB HID mouse driver. Generates a `MouseMove` event
 /// if displacement is non-zero and individual `MouseButton` events
 /// for each of the three standard buttons.
-pub fn inject_mouse_event(dx: i8, dy: i8, buttons: u8) {
+pub fn inject_mouse_event(dx: i16, dy: i16, buttons: u8) {
     let ts = now_ticks();
     static PREV_BUTTONS: spin::Mutex<u8> = spin::Mutex::new(0);
 
     // Displacement event
     if dx != 0 || dy != 0 {
         let event = InputEvent {
-            event_type: InputEventType::MouseMove(dx as i16, dy as i16),
+            event_type: InputEventType::MouseMove(dx, dy),
             timestamp: ts,
         };
         EVENT_QUEUE.lock().push(event);
